indica si el formato de fecha o hora introducidos son correctos

+++ php/borrarTurno.php
@@ -1,22 +1,24 @@
 <?php
     session_start();
     if(isset($_GET['recurso'])){
-        require_once('configuracionDB.php');
-        $conexion=new mysqli(DB_DSN,DB_USUARIO,DB_CONTRASENIA,DB_NAME);
+        if(!empty($_GET['recurso'])){
+            require_once('configuracionDB.php');
+            $conexion=new mysqli(DB_DSN,DB_USUARIO,DB_CONTRASENIA,DB_NAME);
 
-        $sql = "SELECT codigo_recurso FROM " . TABLA_TURNO . " WHERE codigo_recurso = '".$_GET['recurso']."'";
+            $sql = "SELECT codigo_recurso FROM " . TABLA_TURNO . " WHERE codigo_recurso = '".$_GET['recurso']."'";
 
-        //se realiza la consulta
-        if ($resul = $conexion->query($sql)){
-            if($fila = $resul->fetch_row()){
-                //echo "<p>".$fila[0]."</p>";
-                //si se encuentra el turno se borra
-                $sql_delete="DELETE FROM " . TABLA_TURNO . " WHERE codigo_recurso = '".$_GET['recurso']."'";
-                $conexion->query($sql_delete);
+            //se realiza la consulta
+            if ($resul = $conexion->query($sql)){
+                if($fila = $resul->fetch_row()){
+                    //echo "<p>".$fila[0]."</p>";
+                    //si se encuentra el turno se borra
+                    $sql_delete="DELETE FROM " . TABLA_TURNO . " WHERE codigo_recurso = '".$_GET['recurso']."'";
+                    $conexion->query($sql_delete);
+                }
+
             }
-
+            $conexion->close();
         }
-        $conexion->close();
     }
     if($_SESSION['tipo']==0){
         echo "<script language=\"javascript\">window.location=\"../paginaAdmin.php\"</script>;";
